vault: fix vault token upgrade path test
This test requires a task path to be deleted, but on Linux systems it is
mounted in a tmpfs, resulting in an error if the path is deleted before
unmounting.

Since our CI runs on Linux machines, restrict the test to Linux
environments and unmount path before deleting.
